feat: improve tui when generate
