+++ examples/with-secret/kubricate.config.ts
@@ -6,9 +6,6 @@ export default defineConfig({
   stacks: {
     ...simpleAppStack,
   },
-  generate: {
-    outputMode: 'resource',
-  },
   secret: {
     manager: secretManager,
     conflict: {
