+++ packages/kubricate/src/commands/generate/GenerateRunner.ts
@@ -1,13 +1,12 @@
 import fs from 'node:fs';
 import path from 'node:path';
-import { merge } from 'lodash-es';
 import type { BaseLogger, ProjectGenerateOptions } from '@kubricate/core';
+import { MARK_CHECK } from '../../internal/constant.js';
+import type { GlobalConfigOptions } from '../../internal/types.js';
 
-const defaultOptions: Required<ProjectGenerateOptions> = {
-  outputDir: 'output',
-  outputMode: 'stack',
-  cleanOutputDir: true,
-};
+export interface GenerateCommandOptions extends GlobalConfigOptions {
+  outDir: string;
+}
 
 export interface RenderedFile {
   filePath: string;
@@ -15,32 +14,31 @@ export interface RenderedFile {
 }
 
 export class GenerateRunner {
-  public readonly config: Required<ProjectGenerateOptions>;
 
   constructor(
-    config: ProjectGenerateOptions | undefined,
+    public readonly options: GenerateCommandOptions,
+    public readonly generateObjects: Required<ProjectGenerateOptions>,
     private readonly renderedFiles: RenderedFile[],
     protected readonly logger: BaseLogger,
-  ) {
-    this.config = merge({}, defaultOptions, config);
-  }
+  ) { }
 
   async run() {
-    if (this.config.cleanOutputDir) {
-      this.cleanOutputDir(this.config.outputDir);
+    if (this.generateObjects.cleanOutputDir) {
+      this.cleanOutputDir(path.join(this.options.root ?? '', this.generateObjects.outputDir));
     }
 
-    const stats = {
-      written: 0,
-    };
+    const stats = { written: 0 };
 
+    this.logger.info(`Rendering with output mode "${this.generateObjects.outputMode}"\n`);
     for (const { filePath, content } of this.renderedFiles) {
-      this.ensureDir(filePath);
-      fs.writeFileSync(filePath, content);
+      const outputPath = path.join(this.options.root ?? '', filePath);
+      this.ensureDir(outputPath);
+      fs.writeFileSync(outputPath, content);
+      this.logger.log(`${MARK_CHECK} Written file: ${outputPath}`);
       stats.written++;
     }
 
-    this.logger.log(`✅ Wrote ${stats.written} file(s) to "${this.config.outputDir}/"`);
+    this.logger.log(`\n${MARK_CHECK} Written ${stats.written} file${stats.written > 1 ? 's' : ''} to "${this.generateObjects.outputDir}/"`);
   }
 
   private cleanOutputDir(dir: string) {
