+++ packages/kubricate/src/internal/utils.ts
@@ -1,8 +1,8 @@
 import type { GlobalConfigOptions } from './types.js';
 import type { BaseLogger, BaseStack, KubricateConfig, ResourceEntry } from '@kubricate/core';
 import { getMatchConfigFile } from './load-config.js';
-import type { GenerateCommandOptions } from '../commands/generate.js';
 import type { SecretCommandOptions } from '../commands/secret.js';
+import type { GenerateCommandOptions } from '../commands/generate/index.js';
 
 export function getClassName(obj: unknown): string {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
